chore: move is_jwt to src/lib
